APAred {ConsRank}	R Documentation

American Psychological Association dataset, reduced version with only full rankings

Description

The American Psychological Association reduced dataset includes 5738 ballots of the election of the president in 1980, in which the candidates are ranked from most to least favorite.

Usage

data(APAred)

Source

Diaconis, P. (1988). Group representations in probability and statistics. Lecture Notes-Monograph Series, i-192., pag. 96.

[Package ConsRank version 3.0 Index]