kemenydesign {ConsRank}	R Documentation

Auxiliary function

Description

Define a design matrix to compute Kemeny distance

Usage

kemenydesign(X)

Arguments

X
A N by M data matrix, in which there are N judges and M objects to be judged. Each row is a ranking of the objects represented by the columns.

Details

This function now uses an optimized C++ implementation (kemenydesign_cpp) for improved performance. The original R implementation is preserved for reference.

Value

Design matrix

Author(s)

Antonio D'Ambrosio antdambr@unina.it

References

D'Ambrosio, A. (2008). Tree based methods for data editing and preference rankings. Unpublished PhD Thesis. Universita' degli Studi di Napoli Federico II.

See Also

kemenydesign_cpp

[Package ConsRank version 3.0 Index]