kemenyscore {ConsRank}	R Documentation

Score matrix according Kemeny (1962)

Description

Given a ranking, it computes the score matrix as defined by Emond and Mason (2002)

Usage

kemenyscore(X)

Arguments

X
a ranking (must be a row vector or, better, a matrix with one row and M columns)

Value

the M by M score matrix

Author(s)

Antonio D'Ambrosio antdambr@unina.it

References

Kemeny, J and Snell, L. (1962). Mathematical models in the social sciences.

See Also

scorematrix The score matrix as defined by Emond and Mason (2002)

Examples

Y <- matrix(c(1,3,5,4,2),1,5)
SM<-kemenyscore(Y)
#
Z<-c(1,2,3,2)
SM2<-kemenyscore(Z)

[Package ConsRank version 3.0 Index]