order2rank {ConsRank}	R Documentation

Given an ordering, it is transformed to a ranking

Description

From ordering to rank. IMPORTANT: check which symbol denotes tied rankings in the X matrix

Usage

order2rank(X, TO = "{", TC = "}", NAs = "NA")

Arguments

X
A ordering or a matrix containing orderings

TO
symbol indicating the start of a set of items ranked in a tie

TC
symbol indicating the end of a set of items ranked in a tie

NAs
symbol indicating missing ranked object. Can be any symbol in the original ordering matrix

Value

a ranking or a matrix of rankings:

R 	 	 ranking or matrix of rankings

Author(s)

Antonio D'Ambrosio antdambr@unina.it

Examples

data(APAred)
ord=rank2order(APAred) #transform rankings into orderings
ran=order2rank(ord) #transform the orderings into rankings

[Package ConsRank version 3.0 Index]