polyplot {ConsRank}	R Documentation

Plot rankings on a permutation polytope of 3 o 4 objects containing all possible ties

Description

Plot rankings a permutation polytope that is the geometrical space of preference rankings. The plot is available for 3 or for 4 objects

Usage

polyplot(X = NULL, L = NULL, Wk = NULL, nobj = 3)

Arguments

X
the sample of rankings. Most of the time it is returned by tabulaterows

L
labels of the objects

Wk
frequency associated to each ranking

nobj
number of objects. It must be either 3 or 4

Details

polyplot() plots the universe of 3 objects. polyplot(nobj=4) plots the universe of 4 objects. Note: 4-object plots require the 'rgl' package to be installed.

Value

the permutation polytope

Author(s)

Antonio D'Ambrosio antdambr@unina.it and Sonia Amodio sonia.amodio@unina.it

References

Thompson, G. L. (1993). Generalized permutation polytopes and exploratory graphical methods for ranked data. The Annals of Statistics, 1401-1430. # Heiser, W. J., and D'Ambrosio, A. (2013). Clustering and prediction of rankings within a Kemeny distance framework. In Algorithms from and for Nature and Life (pp. 19-31). Springer International Publishing.

See Also

tabulaterows frequency distribution for ranking data.

Examples

## Not run: 
polyplot()

## End(Not run)

## Not run: 
# Requires rgl package
polyplot(nobj=4)
data(BU)
polyplot(BU[,1:3],Wk=BU[,4])

## End(Not run)

[Package ConsRank version 3.0 Index]