rank2order {ConsRank}	R Documentation

Given a rank, it is transformed to a ordering

Description

From ranking to ordering. IMPORTANT: check which symbol denotes tied rankings in the X matrix

Usage

rank2order(X, items = NULL, TO = "{", TC = "}", itemtype = "L")

Arguments

X
A ordering or a matrix containing orderings

items
items to be placed into the ordering matrix. Default are the

TO
symbol indicating the start of a set of items ranked in a tie

TC
symbol indicating the end of a set of items ranked in a tie

itemtype
to be used only if items=NULL. The default value is "L", namely

Value

a ordering or a matrix of orderings:

out 	 	 ranking or matrix of rankings

Author(s)

Antonio D'Ambrosio antdambr@unina.it

Examples

data(APAred)
ord<-rank2order(APAred)

[Package ConsRank version 3.0 Index]