tabulaterows {ConsRank}	R Documentation

Frequency distribution of a sample of rankings

Description

Given a sample of preference rankings, it compute the frequency associated to each ranking

Usage

tabulaterows(X, miss = FALSE)

Arguments

X
a N by M data matrix containing N judges judging M objects

miss
TRUE if there are missing data (either partial or incomplete rankings): default: FALSE

Value

a "list" containing the following components:

X 	 	 the unique rankings

Wk 	 	 the frequency associated to each ranking

tabfreq	 	 frequency table

Author(s)

Antonio D'Ambrosio antdambr@unina.it

Examples

data(Idea)
TR<-tabulaterows(Idea)
FR<-TR$Wk/sum(TR$Wk)
RF<-cbind(TR$X,FR)
colnames(RF)<-c(colnames(Idea),"fi")
#compute modal ranking
maxfreq<-which(RF[,6]==max(RF[,6]))
rank2order(RF[maxfreq,1:5],items=colnames(Idea))
#
data(APAred)
TR<-tabulaterows(APAred)
#
data(APAFULL)
TR<-tabulaterows(APAFULL)
CR1<-consrank(TR$X,wk=TR$Wk)
CR2<-consrank(TR$X,wk=TR$Wk,algorithm="fast",itermax=15)
CR3<-consrank(TR$X,wk=TR$Wk,algorithm="quick")

[Package ConsRank version 3.0 Index]