tau_x {ConsRank}	R Documentation

TauX (tau exstension) rank correlation coefficient

Description

Tau exstension is a new rank correlation coefficient defined by Emond and Mason (2002)

Usage

tau_x(X, Y = NULL)

Tau_X(X, Y = NULL)

Arguments

X
a M by N data matrix, in which there are N judges and M objects to be judged. Each row is a ranking of the objects which are represented by the columns. If there is only X as input, the output is a square matrix containing the Tau_X rcc.

Y
A row vector, or a n by M data matrix in which there are n judges and the same M objects as X to be judged.

Value

Tau_x rank correlation coefficient

Author(s)

Antonio D'Ambrosio antdambr@unina.it

References

Emond, E. J., and Mason, D. W. (2002). A new rank correlation coefficient with application to the consensus ranking problem. Journal of Multi-Criteria Decision Analysis, 11(1), 17-28.

See Also

kemenyd Kemeny distance

iw_tau_x item-weighted tau_x rank correlation coefficient

Examples

data(BU)
RD<-BU[,1:3]
Tau<-tau_x(RD)
Tau1_3<-tau_x(RD[1,],RD[3,])

[Package ConsRank version 3.0 Index]